Moved camera.cc/.h to hydra folder
