+++ source/blender/render/hydra/camera.h
