+++ source/blender/render/hydra/finalEngine.cc
@@ -11,7 +11,7 @@
 #include "glog/logging.h"
 
 #include "finalEngine.h"
-#include "sceneDelegate/camera.h"
+#include "camera.h"
 #include "utils.h"
 
 using namespace std;
